requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mangum>=0.17.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
alembic>=1.12.0

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9

# AWS
boto3>=1.34.0
aws-lambda-powertools>=2.15.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP & Networking
requests>=2.31.0
httpx>=0.25.0

# Data Processing
pandas>=2.2.3
openpyxl>=3.1.0

# Monitoring & Logging
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Security Scanning
bandit>=1.7.5
safety>=2.3.0